Moved the OtherUser interface to the user interfaces file.
